Start separated out dictionary utils
