+++ src/web/client/pages/dictionary_utils.tsx
@@ -1,0 +1,164 @@
+import React, { MutableRefObject } from "react";
+
+import { checkPresent } from "@/common/assert";
+import { XmlNode } from "@/common/xml_node";
+import {
+  ClickableTooltip,
+  SectionLinkTooltip,
+} from "@/web/client/pages/tooltips";
+
+export const SCROLL_JUMP: ScrollIntoViewOptions = {
+  behavior: "auto",
+  block: "start",
+};
+export const SCROLL_SMOOTH: ScrollIntoViewOptions = {
+  behavior: "smooth",
+  block: "start",
+};
+export const ERROR_MESSAGE = {
+  entry: new XmlNode(
+    "span",
+    [],
+    ["Failed to fetch the entry. Please try again later."]
+  ),
+};
+const HIGHLIGHT_HELP = new XmlNode(
+  "div",
+  [["className", "lsHelpText"]],
+  [
+    "Click on ",
+    new XmlNode(
+      "span",
+      [
+        ["class", "lsHover"],
+        ["title", "Click to dismiss"],
+      ],
+      ["underlined"]
+    ),
+    " text for more details. ",
+  ]
+);
+const BUG_HELP = new XmlNode(
+  "div",
+  // This is displayed last, so do not apply the help text style to
+  // avoid an extra margin. Yes, this is gross and hacky.
+  [],
+  [
+    "Please report typos or other bugs " +
+      "by clicking on the flag icon in the top bar.",
+  ]
+);
+const BULLET_HELP = new XmlNode(
+  "div",
+  [["className", "lsHelpText"]],
+  [
+    "Click on sections (like ",
+    new XmlNode(
+      "span",
+      [
+        ["class", "lsSenseBullet"],
+        ["senseid", "tutorialExample"],
+      ],
+      [" A. "]
+    ),
+    ") to link directly to that section.",
+  ]
+);
+export const HELP_ENTRY = new XmlNode(
+  "div",
+  [],
+  [HIGHLIGHT_HELP, BULLET_HELP, BUG_HELP]
+);
+
+export const LOADING_ENTRY = xmlNodeToJsx(
+  new XmlNode(
+    "div",
+    [],
+    [
+      "Please wait - checking for results." +
+        "Dedit oscula nato non iterum repetenda suo ".repeat(3),
+    ]
+  )
+);
+
+export function SelfLink(props: { to: string }) {
+  return <a href={props.to}>{props.to}</a>;
+}
+
+export function xmlNodeToJsx(
+  root: XmlNode,
+  highlightId?: string,
+  sectionRef?: MutableRefObject<HTMLElement | null>,
+  key?: string
+): JSX.Element {
+  const children = root.children.map((child, i) => {
+    if (typeof child === "string") {
+      return child;
+    }
+    return xmlNodeToJsx(
+      child,
+      highlightId,
+      sectionRef,
+      child.getAttr("id") || `${i}`
+    );
+  });
+  const props: { [propKey: string]: any } = {};
+  if (key !== undefined) {
+    props.key = key;
+  }
+  let titleText: string | undefined = undefined;
+  let className: string | undefined = undefined;
+  for (const [attrKey, value] of root.attrs) {
+    if (attrKey === "class") {
+      className = value;
+      props.className = value;
+      continue;
+    }
+    if (attrKey === "title") {
+      titleText = value;
+      continue;
+    }
+    props[attrKey] = value;
+  }
+
+  if (titleText !== undefined) {
+    function hoverForwardedNode(forwardProps: any, forwardRef: any) {
+      const allProps = { ...props, ...forwardProps };
+      allProps["ref"] = forwardRef;
+      return React.createElement(root.name, allProps, children);
+    }
+    const ForwardedNode = React.forwardRef<HTMLElement>(hoverForwardedNode);
+    return (
+      <ClickableTooltip
+        titleText={titleText}
+        className={className}
+        ChildFactory={ForwardedNode}
+        key={key}
+      />
+    );
+  } else if (className === "lsSenseBullet") {
+    function senseForwardedNode(forwardProps: any, forwardRef: any) {
+      const allProps = { ...props, ...forwardProps };
+      allProps["ref"] = forwardRef;
+      return React.createElement(root.name, allProps, children);
+    }
+    const ForwardedNode = React.forwardRef<HTMLElement>(senseForwardedNode);
+    return (
+      <SectionLinkTooltip
+        forwarded={ForwardedNode}
+        className={className}
+        senseId={checkPresent(
+          root.getAttr("senseid"),
+          "lsSenseBullet must have senseid!"
+        )}
+        key={key}
+      />
+    );
+  } else {
+    if (root.getAttr("id") === highlightId && highlightId !== undefined) {
+      props["className"] = "highlighted";
+      props["ref"] = sectionRef!;
+    }
+    return React.createElement(root.name, props, children);
+  }
+}
